chore(DRM): Move util methods from DrmEngine to DrmUtils
