+++ lib/hls/hls_parser.js
@@ -16,7 +16,6 @@ goog.require('shaka.hls.PlaylistType');
 goog.require('shaka.hls.Tag');
 goog.require('shaka.hls.Utils');
 goog.require('shaka.log');
-goog.require('shaka.media.DrmEngine');
 goog.require('shaka.media.InitSegmentReference');
 goog.require('shaka.media.ManifestParser');
 goog.require('shaka.media.PresentationTimeline');
@@ -27,6 +26,7 @@ goog.require('shaka.net.DataUriPlugin');
 goog.require('shaka.net.NetworkingEngine');
 goog.require('shaka.util.ArrayUtils');
 goog.require('shaka.util.BufferUtils');
+goog.require('shaka.util.DrmUtils');
 goog.require('shaka.util.ContentSteeringManager');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.EventManager');
@@ -1936,7 +1936,7 @@ shaka.hls.HlsParser = class {
       audioInfos, videoInfos, bandwidth, width, height, frameRate, videoRange,
       videoLayout, drmInfos, keyIds) {
     const ContentType = shaka.util.ManifestParserUtils.ContentType;
-    const DrmEngine = shaka.media.DrmEngine;
+    const DrmUtils = shaka.util.DrmUtils;
 
     for (const info of videoInfos) {
       this.addVideoAttributes_(
@@ -1979,7 +1979,7 @@ shaka.hls.HlsParser = class {
         const variantUriKey = videoStreamUri + ' - ' + audioStreamUri;
 
         if (audioStream && videoStream) {
-          if (!DrmEngine.areDrmCompatible(audioDrmInfos, videoDrmInfos)) {
+          if (!DrmUtils.areDrmCompatible(audioDrmInfos, videoDrmInfos)) {
             shaka.log.warning(
                 'Incompatible DRM info in HLS variant.  Skipping.');
             continue;
