+++ lib/media/drm_engine.js
@@ -18,7 +18,6 @@ goog.require('shaka.util.EventManager');
 goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.IDestroyable');
 goog.require('shaka.util.Iterables');
-goog.require('shaka.util.Lazy');
 goog.require('shaka.util.ManifestParserUtils');
 goog.require('shaka.util.MapUtils');
 goog.require('shaka.util.MimeUtils');
@@ -760,13 +759,6 @@ shaka.media.DrmEngine = class {
   /** @return {boolean} */
   initialized() {
     return this.initialized_;
-  }
-
-  /**
-   * @param {?shaka.extern.DrmInfo} drmInfo
-   * @return {string} */
-  static keySystem(drmInfo) {
-    return drmInfo ? drmInfo.keySystem : '';
   }
 
   /**
@@ -1764,29 +1756,12 @@ shaka.media.DrmEngine = class {
   }
 
   /**
-   * Returns true if the browser has recent EME APIs.
-   *
-   * @return {boolean}
-   */
-  static isBrowserSupported() {
-    const basic =
-        !!window.MediaKeys &&
-        !!window.navigator &&
-        !!window.navigator.requestMediaKeySystemAccess &&
-        !!window.MediaKeySystemAccess &&
-        // eslint-disable-next-line no-restricted-syntax
-        !!window.MediaKeySystemAccess.prototype.getConfiguration;
-
-    return basic;
-  }
-
-  /**
    * Returns a Promise to a map of EME support for well-known key systems.
    *
    * @return {!Promise.<!Object.<string, ?shaka.extern.DrmSupportType>>}
    */
   static async probeSupport() {
-    goog.asserts.assert(shaka.media.DrmEngine.isBrowserSupported(),
+    goog.asserts.assert(shaka.util.DrmUtils.isBrowserSupported(),
         'Must have basic EME support');
 
     const testKeySystems = [
@@ -2115,89 +2090,11 @@ shaka.media.DrmEngine = class {
       }
     }
 
-    const keySystem = shaka.media.DrmEngine.keySystem(this.currentDrmInfo_);
+    const keySystem = shaka.util.DrmUtils.keySystem(this.currentDrmInfo_);
     const drmInfos = this.getVariantDrmInfos_(variant);
 
     return drmInfos.length == 0 ||
         drmInfos.some((drmInfo) => drmInfo.keySystem == keySystem);
-  }
-
-  /**
-   * Checks if two DrmInfos can be decrypted using the same key system.
-   * Clear content is considered compatible with every key system.
-   *
-   * @param {!Array.<!shaka.extern.DrmInfo>} drms1
-   * @param {!Array.<!shaka.extern.DrmInfo>} drms2
-   * @return {boolean}
-   */
-  static areDrmCompatible(drms1, drms2) {
-    if (!drms1.length || !drms2.length) {
-      return true;
-    }
-
-    if (drms1 === drms2) {
-      return true;
-    }
-
-    return shaka.media.DrmEngine.getCommonDrmInfos(
-        drms1, drms2).length > 0;
-  }
-
-  /**
-   * Returns an array of drm infos that are present in both input arrays.
-   * If one of the arrays is empty, returns the other one since clear
-   * content is considered compatible with every drm info.
-   *
-   * @param {!Array.<!shaka.extern.DrmInfo>} drms1
-   * @param {!Array.<!shaka.extern.DrmInfo>} drms2
-   * @return {!Array.<!shaka.extern.DrmInfo>}
-   */
-  static getCommonDrmInfos(drms1, drms2) {
-    if (!drms1.length) {
-      return drms2;
-    }
-    if (!drms2.length) {
-      return drms1;
-    }
-
-    const commonDrms = [];
-
-    for (const drm1 of drms1) {
-      for (const drm2 of drms2) {
-        if (drm1.keySystem == drm2.keySystem) {
-          const initDataMap = new Map();
-          const bothInitDatas = (drm1.initData || [])
-              .concat(drm2.initData || []);
-          for (const d of bothInitDatas) {
-            initDataMap.set(d.keyId, d);
-          }
-          const initData = Array.from(initDataMap.values());
-
-          const keyIds = drm1.keyIds && drm2.keyIds ?
-              new Set([...drm1.keyIds, ...drm2.keyIds]) :
-              drm1.keyIds || drm2.keyIds;
-          const mergedDrm = {
-            keySystem: drm1.keySystem,
-            licenseServerUri: drm1.licenseServerUri || drm2.licenseServerUri,
-            distinctiveIdentifierRequired: drm1.distinctiveIdentifierRequired ||
-                drm2.distinctiveIdentifierRequired,
-            persistentStateRequired: drm1.persistentStateRequired ||
-                drm2.persistentStateRequired,
-            videoRobustness: drm1.videoRobustness || drm2.videoRobustness,
-            audioRobustness: drm1.audioRobustness || drm2.audioRobustness,
-            serverCertificate: drm1.serverCertificate || drm2.serverCertificate,
-            serverCertificateUri: drm1.serverCertificateUri ||
-                drm2.serverCertificateUri,
-            initData,
-            keyIds,
-          };
-          commonDrms.push(mergedDrm);
-          break;
-        }
-      }
-    }
-
-    return commonDrms;
   }
 
   /**
@@ -2717,73 +2614,6 @@ shaka.media.DrmEngine = class {
     this.newInitData('cenc', combinedData);
     return this.allSessionsLoaded_;
   }
-
-  /**
-   * A method for generating a key for the MediaKeySystemAccessRequests cache.
-   *
-   * @param {string} videoCodec
-   * @param {string} audioCodec
-   * @param {string} keySystem
-   * @return {string}
-   * @private
-   */
-  static generateKeySystemCacheKey_(videoCodec, audioCodec, keySystem) {
-    return `${videoCodec}#${audioCodec}#${keySystem}`;
-  }
-
-  /**
-   * Check does MediaKeySystemAccess cache contains something for following
-   * attributes.
-   *
-   * @param {string} videoCodec
-   * @param {string} audioCodec
-   * @param {string} keySystem
-   * @return {boolean}
-   */
-  static hasMediaKeySystemAccess(videoCodec, audioCodec, keySystem) {
-    const DrmEngine = shaka.media.DrmEngine;
-    const key = DrmEngine.generateKeySystemCacheKey_(
-        videoCodec, audioCodec, keySystem);
-    return DrmEngine.memoizedMediaKeySystemAccessRequests_.has(key);
-  }
-
-  /**
-   * Get MediaKeySystemAccess object for following attributes.
-   *
-   * @param {string} videoCodec
-   * @param {string} audioCodec
-   * @param {string} keySystem
-   * @return {?MediaKeySystemAccess}
-   */
-  static getMediaKeySystemAccess(videoCodec, audioCodec, keySystem) {
-    const DrmEngine = shaka.media.DrmEngine;
-    const key = DrmEngine.generateKeySystemCacheKey_(
-        videoCodec, audioCodec, keySystem);
-    return DrmEngine.memoizedMediaKeySystemAccessRequests_.get(key) || null;
-  }
-
-  /**
-   * Store MediaKeySystemAccess object associated with specified attributes.
-   *
-   * @param {string} videoCodec
-   * @param {string} audioCodec
-   * @param {string} keySystem
-   * @param {!MediaKeySystemAccess} mksa
-   */
-  static setMediaKeySystemAccess(videoCodec, audioCodec, keySystem, mksa) {
-    const DrmEngine = shaka.media.DrmEngine;
-    const key = DrmEngine.generateKeySystemCacheKey_(
-        videoCodec, audioCodec, keySystem);
-    return DrmEngine.memoizedMediaKeySystemAccessRequests_.set(key, mksa);
-  }
-
-  /**
-   * Clears underlying cache.
-   */
-  static clearMediaKeySystemAccessMap() {
-    const DrmEngine = shaka.media.DrmEngine;
-    DrmEngine.memoizedMediaKeySystemAccessRequests_.clear();
-  }
 };
 
 
@@ -2868,22 +2698,3 @@ shaka.media.DrmEngine.SESSION_LOAD_TIMEOUT_ = 5;
  * @type {number}
  */
 shaka.media.DrmEngine.KEY_STATUS_BATCH_TIME = 0.5;
-
-
-/**
- * Contains the suggested "default" key ID used by EME polyfills that do not
- * have a per-key key status. See w3c/encrypted-media#32.
- * @type {!shaka.util.Lazy.<!ArrayBuffer>}
- */
-shaka.media.DrmEngine.DUMMY_KEY_ID = new shaka.util.Lazy(
-    () => shaka.util.BufferUtils.toArrayBuffer(new Uint8Array([0])));
-
-
-/**
- * A cache that stores the MediaKeySystemAccess result of calling
- * `navigator.requestMediaKeySystemAccess` by a key combination of
- * video/audio codec and key system string.
- *
- * @private {!Map<string, !MediaKeySystemAccess>}
- */
-shaka.media.DrmEngine.memoizedMediaKeySystemAccessRequests_ = new Map();
