+++ lib/player.js
@@ -45,6 +45,7 @@ goog.require('shaka.util.CmcdManager');
 goog.require('shaka.util.CmsdManager');
 goog.require('shaka.util.ConfigUtils');
 goog.require('shaka.util.Dom');
+goog.require('shaka.util.DrmUtils');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.EventManager');
 goog.require('shaka.util.FakeEvent');
@@ -1072,7 +1073,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     // DRM support is not strictly necessary, but the APIs at least need to be
     // there.  Our no-op DRM polyfill should handle that.
     // TODO(#1017): Consider making even DrmEngine optional.
-    const drmSupport = shaka.media.DrmEngine.isBrowserSupported();
+    const drmSupport = shaka.util.DrmUtils.isBrowserSupported();
     if (!drmSupport) {
       return false;
     }
@@ -1469,7 +1470,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       // each playback.
       if (this.config_.streaming.clearDecodingCache) {
         shaka.util.StreamUtils.clearDecodingConfigCache();
-        shaka.media.DrmEngine.clearMediaKeySystemAccessMap();
+        shaka.util.DrmUtils.clearMediaKeySystemAccessMap();
       }
 
       this.manifest_ = null;
@@ -4330,7 +4331,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
    * @export
    */
   keySystem() {
-    return shaka.media.DrmEngine.keySystem(this.drmInfo());
+    return shaka.util.DrmUtils.keySystem(this.drmInfo());
   }
 
   /**
