+++ lib/polyfill/media_capabilities.js
@@ -8,8 +8,8 @@ goog.provide('shaka.polyfill.MediaCapabilities');
 
 goog.require('shaka.log');
 goog.require('shaka.media.Capabilities');
-goog.require('shaka.media.DrmEngine');
 goog.require('shaka.polyfill');
+goog.require('shaka.util.DrmUtils');
 goog.require('shaka.util.Platform');
 
 
@@ -284,14 +284,14 @@ shaka.polyfill.MediaCapabilities = class {
     /** @type {MediaKeySystemAccess} */
     let keySystemAccess = null;
     try {
-      if (shaka.media.DrmEngine.hasMediaKeySystemAccess(
+      if (shaka.util.DrmUtils.hasMediaKeySystemAccess(
           videoCodec, audioCodec, keySystem)) {
-        keySystemAccess = shaka.media.DrmEngine.getMediaKeySystemAccess(
+        keySystemAccess = shaka.util.DrmUtils.getMediaKeySystemAccess(
             videoCodec, audioCodec, keySystem);
       } else {
         keySystemAccess = await navigator.requestMediaKeySystemAccess(
             mcapKeySystemConfig.keySystem, [mediaKeySystemConfig]);
-        shaka.media.DrmEngine.setMediaKeySystemAccess(
+        shaka.util.DrmUtils.setMediaKeySystemAccess(
             videoCodec, audioCodec, keySystem, keySystemAccess);
       }
     } catch (e) {
