+++ lib/polyfill/patchedmediakeys_webkit.js
@@ -8,9 +8,9 @@ goog.provide('shaka.polyfill.PatchedMediaKeysWebkit');
 
 goog.require('goog.asserts');
 goog.require('shaka.log');
-goog.require('shaka.media.DrmEngine');
 goog.require('shaka.polyfill');
 goog.require('shaka.util.BufferUtils');
+goog.require('shaka.util.DrmUtils');
 goog.require('shaka.util.EventManager');
 goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.FakeEventTarget');
@@ -924,7 +924,7 @@ shaka.polyfill.PatchedMediaKeysWebkit.MediaKeyStatusMap = class {
   /** @override */
   forEach(fn) {
     if (this.status_) {
-      fn(this.status_, shaka.media.DrmEngine.DUMMY_KEY_ID.value());
+      fn(this.status_, shaka.util.DrmUtils.DUMMY_KEY_ID.value());
     }
   }
 
@@ -938,7 +938,7 @@ shaka.polyfill.PatchedMediaKeysWebkit.MediaKeyStatusMap = class {
 
   /** @override */
   has(keyId) {
-    const fakeKeyId = shaka.media.DrmEngine.DUMMY_KEY_ID.value();
+    const fakeKeyId = shaka.util.DrmUtils.DUMMY_KEY_ID.value();
     if (this.status_ && shaka.util.BufferUtils.equal(keyId, fakeKeyId)) {
       return true;
     }
