+++ lib/util/periods.js
@@ -8,9 +8,9 @@ goog.provide('shaka.util.PeriodCombiner');
 
 goog.require('goog.asserts');
 goog.require('shaka.log');
-goog.require('shaka.media.DrmEngine');
 goog.require('shaka.media.MetaSegmentIndex');
 goog.require('shaka.media.SegmentIndex');
+goog.require('shaka.util.DrmUtils');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.IReleasable');
 goog.require('shaka.util.LanguageUtils');
@@ -282,7 +282,7 @@ shaka.util.PeriodCombiner = class {
     } else {
       for (const audio of this.audioStreams_) {
         for (const video of this.videoStreams_) {
-          const commonDrmInfos = shaka.media.DrmEngine.getCommonDrmInfos(
+          const commonDrmInfos = shaka.util.DrmUtils.getCommonDrmInfos(
               audio.drmInfos, video.drmInfos);
 
           if (audio.drmInfos.length && video.drmInfos.length &&
@@ -855,7 +855,7 @@ shaka.util.PeriodCombiner = class {
       output.originalId += ',' + (input.originalId || '');
     }
 
-    const commonDrmInfos = shaka.media.DrmEngine.getCommonDrmInfos(
+    const commonDrmInfos = shaka.util.DrmUtils.getCommonDrmInfos(
         output.drmInfos, input.drmInfos);
     if (input.drmInfos.length && output.drmInfos.length &&
         !commonDrmInfos.length) {
@@ -1115,7 +1115,7 @@ shaka.util.PeriodCombiner = class {
     if (outputStream.drmInfos) {
       // Check for compatible DRM systems.  Note that clear streams are
       // implicitly compatible with any DRM and with each other.
-      if (!shaka.media.DrmEngine.areDrmCompatible(outputStream.drmInfos,
+      if (!shaka.util.DrmUtils.areDrmCompatible(outputStream.drmInfos,
           candidate.drmInfos)) {
         return false;
       }
