+++ test/media/drm_engine_unit.js
@@ -187,7 +187,7 @@ describe('DrmEngine', () => {
       const variants = manifest.variants;
       await drmEngine.initForPlayback(variants, manifest.offlineSessionIds);
       expect(drmEngine.initialized()).toBe(true);
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo()))
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo()))
           .toBe('drm.abc');
     });
 
@@ -216,7 +216,7 @@ describe('DrmEngine', () => {
       const variants = manifest.variants;
       await drmEngine.initForPlayback(variants, manifest.offlineSessionIds);
       expect(drmEngine.initialized()).toBe(true);
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo()))
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo()))
           .toBe('drm.def');
     });
 
@@ -233,7 +233,7 @@ describe('DrmEngine', () => {
       // should be only one variant, as preferredKeySystems is propagated
       // to getDecodingInfos
       expect(variants[0].decodingInfos.length).toBe(1);
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo()))
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo()))
           .toBe('drm.def');
     });
 
@@ -251,7 +251,7 @@ describe('DrmEngine', () => {
       await drmEngine.initForPlayback(variants, manifest.offlineSessionIds);
 
       expect(variants[0].decodingInfos.length).toBe(2);
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo()))
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo()))
           .toBe('drm.def');
     });
 
@@ -524,7 +524,7 @@ describe('DrmEngine', () => {
           expect(variants[0].decodingInfos.length).toBe(1);
           expect(variants[0].decodingInfos[0].keySystemAccess).toBeFalsy();
           expect(
-              shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo())).toBe('');
+              shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo())).toBe('');
           expect(drmEngine.initialized()).toBe(true);
         });
 
@@ -540,7 +540,7 @@ describe('DrmEngine', () => {
       const variants = manifest.variants;
       await drmEngine.initForPlayback(variants, manifest.offlineSessionIds);
       expect(drmEngine.initialized()).toBe(true);
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo()))
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo()))
           .toBe('drm.abc');
       expect(variants[0].decodingInfos.length).toBe(1);
     });
@@ -1948,7 +1948,7 @@ describe('DrmEngine', () => {
       p.resolve();  // Success for drm.abc.
       await expectAsync(init).toBeRejected();
       // Due to the interruption, we never created MediaKeys.
-      expect(shaka.media.DrmEngine.keySystem(drmEngine.getDrmInfo())).toBe('');
+      expect(shaka.util.DrmUtils.keySystem(drmEngine.getDrmInfo())).toBe('');
       expect(drmEngine.initialized()).toBe(false);
     });
 
@@ -2280,144 +2280,6 @@ describe('DrmEngine', () => {
       });
     });
   });  // describe('getDrmInfo')
-
-  describe('getCommonDrmInfos', () => {
-    it('returns one array if the other is empty', () => {
-      const drmInfo = {
-        keySystem: 'drm.abc',
-        licenseServerUri: 'http://abc.drm/license',
-        distinctiveIdentifierRequired: true,
-        persistentStateRequired: true,
-        audioRobustness: 'good',
-        videoRobustness: 'really_really_ridiculously_good',
-        serverCertificate: undefined,
-        serverCertificateUri: '',
-        initData: [],
-        keyIds: new Set(['deadbeefdeadbeefdeadbeefdeadbeef']),
-      };
-      const returnedOne =
-          shaka.media.DrmEngine.getCommonDrmInfos([drmInfo], []);
-      const returnedTwo =
-          shaka.media.DrmEngine.getCommonDrmInfos([], [drmInfo]);
-      expect(returnedOne).toEqual([drmInfo]);
-      expect(returnedTwo).toEqual([drmInfo]);
-    });
-
-    it('merges drmInfos if two exist', () => {
-      const serverCert = new Uint8Array(0);
-      const drmInfoVideo = {
-        keySystem: 'drm.abc',
-        licenseServerUri: 'http://abc.drm/license',
-        distinctiveIdentifierRequired: false,
-        persistentStateRequired: true,
-        videoRobustness: 'really_really_ridiculously_good',
-        serverCertificate: serverCert,
-        serverCertificateUri: '',
-        initData: [{keyId: 'a'}],
-        keyIds: new Set(['deadbeefdeadbeefdeadbeefdeadbeef']),
-      };
-      const drmInfoAudio = {
-        keySystem: 'drm.abc',
-        licenseServerUri: undefined,
-        distinctiveIdentifierRequired: true,
-        persistentStateRequired: false,
-        audioRobustness: 'good',
-        serverCertificate: undefined,
-        serverCertificateUri: '',
-        initData: [{keyId: 'b'}],
-        keyIds: new Set(['eadbeefdeadbeefdeadbeefdeadbeefd']),
-      };
-      const drmInfoDesired = {
-        keySystem: 'drm.abc',
-        licenseServerUri: 'http://abc.drm/license',
-        distinctiveIdentifierRequired: true,
-        persistentStateRequired: true,
-        audioRobustness: 'good',
-        videoRobustness: 'really_really_ridiculously_good',
-        serverCertificate: serverCert,
-        serverCertificateUri: '',
-        initData: [{keyId: 'a'}, {keyId: 'b'}],
-        keyIds: new Set([
-          'deadbeefdeadbeefdeadbeefdeadbeef',
-          'eadbeefdeadbeefdeadbeefdeadbeefd',
-        ]),
-      };
-      const returned = shaka.media.DrmEngine.getCommonDrmInfos([drmInfoVideo],
-          [drmInfoAudio]);
-      expect(returned).toEqual([drmInfoDesired]);
-    });
-
-    it('dedupes the merged init data based on keyId matching', () => {
-      const serverCert = new Uint8Array(0);
-      const drmInfoVideo = {
-        keySystem: 'drm.abc',
-        licenseServerUri: 'http://abc.drm/license',
-        distinctiveIdentifierRequired: false,
-        persistentStateRequired: true,
-        videoRobustness: 'really_really_ridiculously_good',
-        serverCertificate: serverCert,
-        serverCertificateUri: '',
-        initData: [{keyId: 'v-init'}],
-        keyIds: new Set(['deadbeefdeadbeefdeadbeefdeadbeef']),
-      };
-      const drmInfoAudio = {
-        keySystem: 'drm.abc',
-        licenseServerUri: undefined,
-        distinctiveIdentifierRequired: true,
-        persistentStateRequired: false,
-        audioRobustness: 'good',
-        serverCertificate: undefined,
-        serverCertificateUri: '',
-        initData: [{keyId: 'v-init'}, {keyId: 'a-init'}],
-        keyIds: new Set(['eadbeefdeadbeefdeadbeefdeadbeefd']),
-      };
-      const drmInfoDesired = {
-        keySystem: 'drm.abc',
-        licenseServerUri: 'http://abc.drm/license',
-        distinctiveIdentifierRequired: true,
-        persistentStateRequired: true,
-        audioRobustness: 'good',
-        videoRobustness: 'really_really_ridiculously_good',
-        serverCertificate: serverCert,
-        serverCertificateUri: '',
-        initData: [{keyId: 'v-init'}, {keyId: 'a-init'}],
-        keyIds: new Set([
-          'deadbeefdeadbeefdeadbeefdeadbeef',
-          'eadbeefdeadbeefdeadbeefdeadbeefd',
-        ]),
-      };
-      const returned = shaka.media.DrmEngine.getCommonDrmInfos([drmInfoVideo],
-          [drmInfoAudio]);
-      expect(returned).toEqual([drmInfoDesired]);
-    });
-
-    it('does not match incompatible drmInfos', () => {
-      // Different key systems do not match.
-      const drmInfo1 = {
-        keySystem: 'drm.abc',
-        licenseServerUri: undefined,
-        distinctiveIdentifierRequired: false,
-        persistentStateRequired: false,
-        serverCertificate: undefined,
-        serverCertificateUri: '',
-        initData: [],
-        keyIds: new Set(),
-      };
-      const drmInfo2 = {
-        keySystem: 'drm.foobar',
-        licenseServerUri: undefined,
-        distinctiveIdentifierRequired: false,
-        persistentStateRequired: false,
-        serverCertificate: undefined,
-        serverCertificateUri: '',
-        initData: [],
-        keyIds: new Set(),
-      };
-      const returned1 = shaka.media.DrmEngine.getCommonDrmInfos(
-          [drmInfo1], [drmInfo2]);
-      expect(returned1).toEqual([]);
-    });
-  }); // describe('getCommonDrmInfos')
 
   describe('configure', () => {
     it('delays initial license requests if configured to', async () => {
